[SPARK-53365][SQL] Unify code for persisting of configs in views and UDFs
### What changes were proposed in this pull request?
In this issue I propose that we unify code for persisting the configs in views and UDFs as it should be the same to avoid further discrepancies (until now, we had separated methods and that's why we didn't have SESSION_LOCAL_TIMEZONE stored for UDFs).

### Why are the changes needed?
To improve the code health and fix existing issues.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
Tests added in the PR.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #52113 from mihailoale-db/conftrait.

Authored-by: mihailoale-db <mihailo.aleksic@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>