[SPARK-53398][SS] Ensure that RocksDBMemoryManager metrics reporting doesn't throw a NullPointerException
### What changes were proposed in this pull request?

Appending a UUID to the uniqueId that RocksDB uses to register itself with the memory manager that will prevent a race condition that triggers an NPE and fails the query

### Why are the changes needed?

If the RocksDB instance is unregistered from the memory manager via maintenance, then metrics are called on the same instance, then we will throw an NPE, which will fail the query, from the RocksDBMemoryManager metrics reporting path. We can handle this by appending a UUID to the uniqueId we pass to RocksDB and using getOrDefault in the Memory Manager.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Unit test

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #52144 from ericm-db/metrics-npe.

Authored-by: Eric Marnadi <eric.marnadi@databricks.com>
Signed-off-by: Anish Shrigondekar <anish.shrigondekar@databricks.com>