[SPARK-53421][SPARK-53377][SDP] Propagate Logical Plan ID in SDP Analysis
### What changes were proposed in this pull request?

Propagate `LogicalPlan.PLAN_ID_TAG` to the resolved logical plan during SDP analysis so when the whole plan is sent to Spark for analysis, it contains the correct plan id.

### Why are the changes needed?

Spark Connect attaches a plan id to each logical plan. In SDP, we take part of the logical plan and analyze it independently to resolve table references correctly. When this happens, the logical plan id is lost which causes resolution errors when the plan is sent to Spark for complete analysis.

For example, group by and rollup functions would fail with `sql.AnalysisException: [CANNOT_RESOLVE_DATAFRAME_COLUMN] Cannot resolve dataframe column "id". It's probably because of illegal references like df1.select(df2.col("a"))`
```python3
from pyspark.sql.functions import col, sum, count

dp.materialized_view
def groupby_result():
    return spark.read.table("src").groupBy("id").count()
```
This happens because we take the below unresolved logical plan:
```
'Aggregate ['id], ['id, 'count(1) AS count#7]
+- 'UnresolvedRelation [src], [], false
```
Perform independent analysis on the `UnresolvedRelation` part to identify the table. During this analysis, the plan id is lost.
```
'Aggregate ['id], ['id, 'count(1) AS count#7]
+- SubqueryAlias spark_catalog.default.src
   +- Relation spark_catalog.default.src[id#9L] parquet
```
So when the above partially resolved logical plan is sent to Spark for analysis, it tries to resolve the `id` attribute in the aggregate operation with respect to the `SubqueryAlias` children, and fails because the children no longer contains the same plan id.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Tests

### Was this patch authored or co-authored using generative AI tooling?

Closes #52121 from JiaqiWang18/SPARK-53377-sdp-groupBy-rollup-tests.

Authored-by: Jacky Wang <jacky.wang@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>