[SPARK-53450] Nulls are filled unexpectedly after converting hive table scan to logical relation
### What changes were proposed in this pull request?

This PR introduces a configuration `spark.sql.hive.convertMetastoreAsNullable` to set nullable to true for all fields while converting Hive parquet/orc tables to datasources to fix a data-correctness issue.

### Why are the changes needed?

Hive tables are generally schema-on-read, and those Spark-executed Hive DDLs are not always Hive compatible.

We have an online case that produces incorrect values while `convertMetastoreParquet=true`, for a null record, it sometimes returns 5, others 0, and no other values are witnessed yet.

I've tried 4 ways to read the table and data

1) convertMetastoreParquet=true, vectorized  -- negative
2) convertMetastoreParquet=true, row-based  -- negative
3) convertMetastoreParquet=false  -- positive
4) scan on the table path, partition path, or files -- positive

It turns out that the incorrect table schema led to wrong results.

So, in this PR, I added a new configuration to fall back on how Hive handles nullability blindly.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Unfortunately, I've not yet figured out a reproducible unit test to cover this. So, I packed and tested online with our prod data

#### Before this PR
<img width="387" height="195" alt="image" src="https://github.com/user-attachments/assets/b89836c9-f0c2-4beb-8f19-903f761f212a" />

#### After this PR

<img width="387" height="192" alt="image" src="https://github.com/user-attachments/assets/6b305268-581f-4538-b60a-1d4ae5720fff" />

### Was this patch authored or co-authored using generative AI tooling?
no

Closes #52193 from yaooqinn/SPARK-53450.

Authored-by: Kent Yao <yao@apache.org>
Signed-off-by: Kent Yao <yao@apache.org>