[SPARK-52601][SQL] Support primitive types in TransformingEncoder
### What changes were proposed in this pull request?
Support defining TransformingEncoder that has a primitive type as the input type.

### Why are the changes needed?
To support defining TransformingEncoder that has a primitive type as the input type.

This came up for me when using a Scala 3 opaque type around a Long as a timestamp but wating have the encoder encode it as a timestamp. Ideally Spark would have some way of encoding a micro second timestamp without going through a java.sql.Timestamp or java.time.Instant. But this at least makes it possible to achive something similar (but less efficient) by defining a TransformingEncoder that takes a Long and returns a java.sql.Timestamp.

### Does this PR introduce _any_ user-facing change?
Yes, it allows TransformingEncoder to be used in more cases.

### How was this patch tested?
New and existing unit tests.

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #51313 from eejbyfeldt/SPARK-52601.

Authored-by: Emil Ejbyfeldt <emil.ejbyfeldt@choreograph.com>
Signed-off-by: Herman van Hovell <herman@databricks.com>