[SPARK-53498][SDP] Correctly Reference `pyspark/pipelines/cli.py` from `spark-pipelines` Binary
### What changes were proposed in this pull request?
When the `spark-pipelines` binary is run, SDP's `cli.py` is submitted to the configured spark cluster via a `spark-submit` request.

Previously, the file path to retrieve `cli.py` was hard-coded with respect to the resolved `SPARK_HOME`. This is incorrect as the package layout is different depending on how spark was installed (ex. Spark tarball installation vs PyPi pyspark installation).

The proposed change is to resolve the file path as a hard coded string with respect to where the pyspark module source code is, rather than `SPARK_HOME`.

### Why are the changes needed?

### Does this PR introduce _any_ user-facing change?

### How was this patch tested?

### Was this patch authored or co-authored using generative AI tooling?

Closes #52254 from AnishMahto/fix-sdp-cli-pyspark.

Authored-by: anishm-db <anish.mahto@databricks.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>