[SPARK-53558][SQL] Show fully qualified table name including the catalog name in the exception message when the table is not found
Currently, we only display the database name and table name in the error message when the table is not found. We should display the fully qualified table name, including the catalog name, in the exception message when the table is missing.

### What changes were proposed in this pull request?

### Why are the changes needed?

Better error message.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Modified the affected unit tests and verified that the error message has the fully qualified table name when the table is not found.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #52315 from ganeshashree/SPARK-53558.

Authored-by: Ganesha S <ganesha.s@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>