[SPARK-47110][INFRA] Reenble AmmoniteTest tests in Maven builds
### What changes were proposed in this pull request?
This PR aims to reenable `AmmoniteTest` tests in Maven builds using `script` command.
The usage of the command is different from macOS and Linux so this change consider the difference.

### Why are the changes needed?
For better test coverage.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
I created a branch on my repo based on this change which runs Maven tests for macOS and Ubuntu, and I confirmed both tests passed.
* [Test on Ubuntu](https://github.com/sarutak/spark/actions/runs/17779850500)
* [Test on macOS](https://github.com/sarutak/spark/actions/runs/17779850568)

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #52369 from sarutak/reenble-ammonite-tests.

Authored-by: Kousuke Saruta <sarutak@amazon.co.jp>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>