[SPARK-46679][SQL] Fix for SparkUnsupportedOperationException Not found an encoder of the type T, when using Parameterized class
This pr is a rebase/continue of https://github.com/apache/spark/pull/48304/files, which was unexpectedly closed even though the issue is still there.

### What changes were proposed in this pull request?
That pr fixes a bug in JavaTypeInference.encoderFor.  In the parameterizableType case, it overrides the existing map of parameterizable type -> real type.  So any nested parameterizable type will fail.

### Why are the changes needed?
Fix nested parameterization cases for Java Encoder

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
The original pr adds unit test to JavaTypeInferenceSuite and JavaDatasetSuite.  The test from the original pr are simplified a little to make it obvious which one is wrapping.

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #52444 from szehon-ho/parameterized_encoders.

Authored-by: Szehon Ho <szehon.apache@gmail.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>