[SPARK-53734][SQL] Prefer table column over LCA when resolving array index
### What changes were proposed in this pull request?
Prefer table column over LCA when resolving array index.

### Why are the changes needed?
For a query like:

```
SELECT 1 AS col1, col2[col1] FROM VALUES(0, ARRAY(1,2));
```

the output should be (1,1), but current Spark implementation outputs (1,2). This is because `[col1]` is resolved as an LCA instead of being resolved to a column. This is because we never actually resolve `field` of `UnresolvedExtractValue` in `innerResolve`, so the resolution of `field` fails over to the next item in precedence chain, which is LCA.

### Does this PR introduce _any_ user-facing change?
Yes, user now sees the correct result for the impacted query shape

### How was this patch tested?
Added test case for the impacted query.

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #52472 from mihailotim-db/mihailo-timotic_data/array_index_lca_correctness.

Authored-by: Mihailo Timotic <mihailo.timotic@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>