[SPARK-53868][SQL] Only use signature with Expression[]  of `visitAggregateFunction` in V2ExpressionSQBuilder
### What changes were proposed in this pull request?

https://github.com/apache/spark/pull/50143 This PR introduced `visitAggregateFunction` with `inputs: Array[Expression]`, but it's not the only usage of `visitAggregateFunction` for example here
https://github.com/apache/spark/blob/2aa5dcc27840a88e6853ec6caef7ffd8c3916611/sql/catalyst/src/main/java/org/apache/spark/sql/connector/util/V2ExpressionSQLBuilder.java#L134
the old API is still used but it no longer checked if the function is supported.

This means that if some dialect did not support one of `(MIN, MAX, COUNT, SUM, AVG)` it would not be blocked, but as of now all the dialects have them in the `supportedFunctions` so this is not a behavioral change.

### Why are the changes needed?

To unify the API in case in the future if some dialect does not support an aggregate function of `(MIN, MAX, COUNT, SUM, AVG)` it should be blocked.
### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Existing tests.
### Was this patch authored or co-authored using generative AI tooling?

No

Closes #52573 from alekjarmov/block-aggregates.

Lead-authored-by: alekjarmov <alek.jarmov@databricks.com>
Co-authored-by: Alek Jarmov <alekjarmov@Aleks-MacBook-Pro.local>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>