[SPARK-53720][SQL] Simplify extracting Table from DataSourceV2Relation (#52460)
### What changes were proposed in this pull request?

This PR adds a new extractor for `Table` from `DataSourceV2Relation`.

### Why are the changes needed?

As we see over time, `DataSourceV2Relation` continues to evolve and has many args. Frequently, we only need to get the table instance and are not interested in other arguments. Therefore, it makes sense to add a new extractor for this common use case. In particular, I plan to add `TimeTravelSpec` and I don't want to update a ton of places for no good reason.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

This PR relies on existing tests.

### Was this patch authored or co-authored using generative AI tooling?

No.