[SPARK-53908][CONNECT] Fix observations on Spark Connect with plan cache
### What changes were proposed in this pull request?

Fixes observations on Spark Connect with plan cache enabled.

### Why are the changes needed?

The observations on Spark Connect with plan cache enabled has cases that can't see the observed values.

```py
>>> from pyspark.sql.observation import Observation
>>> from pyspark.sql import functions as F
>>>
>>> df = spark.range(10)
>>> observation = Observation()
>>> observed_df = df.observe(observation, F.count(F.lit(1)).alias("cnt"))
>>>
>>> observed_df.schema  # causes to cache
StructType([StructField('id', LongType(), False)])
>>>
>>> observed_df.show()
+---+
| id|
+---+
...
+---+

>>>
>>> observation.get
{}
```

This should be:

```py
>>> spark.conf.set('spark.connect.session.planCache.enabled', False)
...
>>> observation.get
{'cnt': 10}
```

This is because the cached plan by the Analyze request for `observed_df.schema` doesn't register the new `Observation` for the actual execution.

The observations should be registered later.

### Does this PR introduce _any_ user-facing change?

Yes, the observations will be available.

### How was this patch tested?

Modified the related tests.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #52616 from ueshin/issues/SPARK-53908/observations.

Authored-by: Takuya Ueshin <ueshin@databricks.com>
Signed-off-by: Takuya Ueshin <ueshin@databricks.com>