[SPARK-53968][SQL] Store decimal precision loss conf in arithmetic expressions
### What changes were proposed in this pull request?
Currently, arithmetic expressions such as `Add` and `Multiply` use the configuration `spark.sql.decimalOperations.allowPrecisionLoss` to determine their output type when working with decimal values. This approach is problematic because if the expression is transformed or copied, its return type could change depending on the active configuration value.

This issue can happen during view resolution; we can use one value of the config during analysis and different one during query optimization. If a referenced expression changes type and that reference is reused elsewhere in the plan it will trigger a plan validation error.

### Why are the changes needed?
To address this, we should follow a similar approach to what was done for ANSI mode: store the relevant context directly within the expression as part of its state. This ensures the expression remains stable and unaffected by configuration changes when it’s copied or transformed. To make this transition smooth, I’ve generalized the existing EvalMode used for ANSI so that it can be extended to multiple configuration dimensions.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
Added a new unit test in SQLViewSuite which was failing with plan validation error before.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #52681 from stefankandic/fixViewDec.

Authored-by: Stefan Kandic <stefan.kandic@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>