[SPARK-54015][PYTHON] Relax Py4J requirement to py4j>=0.10.9.7,<0.10.9.10
### What changes were proposed in this pull request?

Py4J py4j>=0.10.9.7,<0.10.9.10 is JVM compatible so we can actually use older versions as well (I released those Py4J versions). To make it easier to migrate to higher versions of Spark, this PR proposes the range.

### Why are the changes needed?

To make it easier for legacy users to migrate to higher versions of Spark.

### Does this PR introduce _any_ user-facing change?

Virtually no.

### How was this patch tested?

Manually.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #52721 from HyukjinKwon/SPARK-54015.

Authored-by: Hyukjin Kwon <gurwls223@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>