[SPARK-53335][K8S] Support `spark.kubernetes.driver.annotateExitException`
### What changes were proposed in this pull request?

This PR aims to provide a way for users to annotate the driver pod with exception message when the applications exit with exceptions.

### Why are the changes needed?

For jobs which run on kubernetes there is no native concept of diagnostics (like there is in YARN), which means that for debugging and triaging errors users _must_ go to logs. For jobs which run on YARN this is often not necessary, since the diagnostics contains the root cause reason for failure. Additionally, for platforms which provide automation of failure insights, or make decisions based on failures, there must be a custom solution or deciding why the application failed (e.g. log and stack trace parsing).

We use a similar mechanism as https://github.com/apache/spark/pull/23599 to load custom implementations in order to avoid the dependency on the k8s module from SparkSubmit.

### Does this PR introduce _any_ user-facing change?

Yes, a config, which is defaulted to false.

### How was this patch tested?

Unit tested + verified in production k8s cluster.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #52068 from ForVic/vsunderl/better_kubernetes_diagnostics.

Lead-authored-by: Victor Sunderland <victors@openai.com>
Co-authored-by: ForVic <vsunderland@linkedin.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>