[SPARK-54085][CORE] Fix `initialize` to add `CREATE` option additionally in `DriverRunner`
### What changes were proposed in this pull request?
Fix the usage of Files.writeString in DriverRunner of Standalone Worker. Currently users will get errors below when submitting jobs to standalone cluster with restful api:

```
25/10/30 16:02:59 INFO DriverRunner: Killing driver process!
25/10/30 16:02:59 INFO CommandUtils: Redirection to /Users/ericgao/workspace/open_source/spark/work/driver-20251030160259-0020/stdout closed: Stream closed
25/10/30 16:02:59 WARN Worker: Driver driver-20251030160259-0020 failed with unrecoverable exception: java.nio.file.NoSuchFileException: /Users/ericgao/workspace/open_source/spark/work/driver-20251030160259-0020/stderr
```

### Why are the changes needed?
With only StandardOpenOption.APPEND:
If the target file already exists, the string is appended to the end of the file.
If the target file does not exist, the call throws NoSuchFileException because append requires an existing file.

With StandardOpenOption.CREATE, StandardOpenOption.APPEND:
If the target file does not exist, it will be created, then the string is appended (effectively written as the first content).
If the target file already exists, CREATE is ignored and the string is appended.

With this fix, users will no more get errors like java.nio.file.NoSuchFileException for stderr when submitting jobs to standalone cluster.

### Does this PR introduce _any_ user-facing change?
No, just tiny bug fix.

### How was this patch tested?
Tested manually.

### Was this patch authored or co-authored using generative AI tooling?

Closes #52789 from EricGao888/fix-writeString-in-standalone-driverRunner.

Authored-by: EricGao888 <ericgao.apache@gmail.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>