[SPARK-53959][CONNECT] Throw a client-side error when creating a dataframe from a pandas dataframe with an index but no data
### What changes were proposed in this pull request?

Spark Connect Python client does not throw a proper error when creating a dataframe from a pandas dataframe with a index and empty data.
Generally, spark connect client throws a client-side error `[CANNOT_INFER_EMPTY_SCHEMA] Can not infer schema from an empty dataset`. when creating a dataframe without data, for example via
```
spark.createDataFrame([]).show()
```
or
```
df = pd.DataFrame()
spark.createDataFrame(df).show()
```
or
```
df = pd.DataFrame({"a": []})
spark.createDataFrame(df).show()
```

This does not happen when pandas dataframe has an index but no data, e.g.
```
df = pd.DataFrame(index=range(5))
spark.createDataFrame(df).show()
```
What happens instead is that the dataframe is successfully converted to a LocalRelation on the client, is sent to the server, but the server then throws the following exception: `INTERNAL_ERROR: Input data for LocalRelation does not produce a schema. SQLSTATE: XX000`. XX000 is an internal error sql state and the error is not actionable enough for the user.

This PR fixes this problem by throwing `CANNOT_INFER_EMPTY_SCHEMA` in case the dataframe has rows (because of the index), but does not have columns.

### Why are the changes needed?

Currently the error is thrown as a server-side internal error and the error message is not actionable enough for the user.

### Does this PR introduce _any_ user-facing change?

Creating a spark connect dataframe from an pandas dataframe with an index but no data will now throw a client-side error `[CANNOT_INFER_EMPTY_SCHEMA] Can not infer schema from an empty dataset` instead of the server-side `INTERNAL_ERROR: Input data for LocalRelation does not produce a schema. SQLSTATE: XX000`.

### How was this patch tested?

New unit test.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #52670 from khakhlyuk/pd-empty-data.

Authored-by: Alex Khakhlyuk <alex.khakhlyuk@gmail.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>